Resolves #2121: Refactor KeySpaceTreeResolver out of KeySpaceCountTree
